// AestheticEMR/AestheticEMR.client/src/app/models/legacy/h-retainership.model.ts
export interface HRetainership {
  retainId: string;
  retainCode: string;
  retainName: string;
  clientCatId?: string;
  clientType?: string;
  address?: string;
  phoneNo?: string;
  email?: string;
  contact?: string;
  profFee?: number;
  debt?: number;
  acctId?: string;
  debtType?: string;
  active?: string;
  useTariff?: string;
  pcent?: number;
  billEndDate?: number;
  regAmount?: number;
  conAmount?: number;
  cardRenewAmount?: number;
  retainDate?: string;
  entryDate?: string;
  entryTime?: string;
  clientName?: string;
  appName?: string;
}